fix: keep main green under the #9367/#9372/#9373 batch
- split object/mod.rs test root accessors and object/tests.rs transition-IC
  tests into siblings, both files were within ~35 lines of the 2000 cap
- refresh the shape-descriptor census: one new object_header_size callsite in
  proxy_reflect.rs (42 -> 43) and one keys_array access relocated by the split
- drop a redundant unsafe block that -D warnings rejects
